Refactor and consolidate CE status effects
Consolidate and refactor many CE status effect components into Content.Shared._CE.StatusEffects: moved/renamed multiple component files, added RegisterComponent/NetworkedComponent (and AutoGenerateComponentState where needed), and standardized data fields. Introduced CEEffectOnLandingStatusEffect to run entity effects on landing and hooked relay to CEZLevelFallMapEvent. Added a server-side CEAdditionalRoleSkillsStatusEffectComponent on the JobWhitelist to grant extra role skills during blessing selection. Added new skill prototypes and assets (Heavyweight, LandingTechnique, Jump) and updated related prototype icons and metadata. Removed legacy component files in their old locations as part of the reorganization.
