+++ Content.Shared/StatusEffectNew/StatusEffectSystem.Relay.cs
@@ -1,14 +1,15 @@
 using System.Linq;
 using Content.Shared._CE.DebuffCleaning;
 using Content.Shared._CE.EntityEffect.Effects;
-using Content.Shared._CE.DivineShield;
 using Content.Shared._CE.Health;
 using Content.Shared._CE.Mana.Core;
 using Content.Shared._CE.MeleeWeapon;
 using Content.Shared._CE.Soul;
 using Content.Shared._CE.Stamina;
+using Content.Shared._CE.StatusEffects;
 using Content.Shared._CE.StatusEffects.Core;
 using Content.Shared._CE.TileEffects.Core;
+using Content.Shared._CE.ZLevels.Core.EntitySystems;
 using Content.Shared._CE.ZLevels.Damage;
 using Content.Shared.Actions.Events;
 using Content.Shared.Body.Events;
@@ -72,9 +73,11 @@ public sealed partial class StatusEffectsSystem
         SubscribeLocalEvent<StatusEffectContainerComponent, IsEquippingAttemptEvent>(RelayStatusEffectEvent);
         SubscribeLocalEvent<StatusEffectContainerComponent, IsUnequippingAttemptEvent>(RelayStatusEffectEvent);
         SubscribeLocalEvent<StatusEffectContainerComponent, CEZFallingDamageCalculateEvent>(RelayStatusEffectEvent);
+
         SubscribeLocalEvent<StatusEffectContainerComponent, CESoulReceivedEvent>(RefRelayStatusEffectEvent);
         SubscribeLocalEvent<StatusEffectContainerComponent, CEAttemptApplyTileEffectEvent>(RefRelayStatusEffectEvent);
         SubscribeLocalEvent<StatusEffectContainerComponent, ActionAttemptEvent>(RefRelayStatusEffectEvent);
+        SubscribeLocalEvent<StatusEffectContainerComponent, CEZLevelFallMapEvent>(RefRelayStatusEffectEvent);
 
         //CrystallEdge zone end
