+++ Content.Shared/_CE/StatusEffects/CEFallingImmunityStatusEffect.cs
@@ -1,7 +1,19 @@
 using Content.Shared._CE.ZLevels.Damage;
 using Content.Shared.StatusEffectNew;
+using Robust.Shared.GameStates;
 
-namespace Content.Shared._CE.StatusEffects.FallingImmune;
+namespace Content.Shared._CE.StatusEffects;
+
+[RegisterComponent, NetworkedComponent]
+public sealed partial class CEFallingImmunityStatusEffectComponent : Component
+{
+    [DataField]
+    public float DamageMultiplier = 1f;
+
+    [DataField]
+    public float StunMultiplier = 1f;
+}
+
 
 public sealed class CEFallingImmunityStatusEffectSystem : EntitySystem
 {
