+++ Content.Shared/_CE/StatusEffects/CEIncomingHealBonusStatusEffect.cs
@@ -1,8 +1,20 @@
 using Content.Shared._CE.Health;
 using Content.Shared._CE.StatusEffects.Core.Components;
 using Content.Shared.StatusEffectNew;
+using Robust.Shared.GameStates;
 
-namespace Content.Shared._CE.StatusEffects.IncomingHealBonus;
+namespace Content.Shared._CE.StatusEffects;
+
+/// <summary>
+/// Status effect that increases incoming healing on the entity it's applied to.
+/// Scales with stacks via <see cref="CEStatusEffectStackComponent"/>.
+/// </summary>
+[RegisterComponent, NetworkedComponent]
+public sealed partial class CEIncomingHealBonusStatusEffectComponent : Component
+{
+    [DataField]
+    public int BonusPerStack = 1;
+}
 
 public sealed partial class CEIncomingHealBonusStatusEffectSystem : EntitySystem
 {
