+++ Content.Shared/_CE/StatusEffects/CEPermanentStatusEffects.cs
@@ -1,6 +1,19 @@
 using Content.Shared.StatusEffectNew;
+using Robust.Shared.GameStates;
+using Robust.Shared.Prototypes;
 
 namespace Content.Shared._CE.StatusEffects;
+
+/// <summary>
+/// Adds permanent status effects to the entity when the component is initialized.
+/// </summary>
+[RegisterComponent, NetworkedComponent]
+public sealed partial class CEPermanentStatusEffectsComponent : Component
+{
+    [DataField(required: true)]
+    public HashSet<EntProtoId> Effects = new();
+}
+
 
 public sealed partial class CEPermanentStatusEffectsSystem : EntitySystem
 {
