--- Content.Shared/_CE/StatusEffects/CEPermanentStatusEffectsComponent.cs
@@ -1,11 +1,0 @@
-using Robust.Shared.GameStates;
-using Robust.Shared.Prototypes;
-
-namespace Content.Shared._CE.StatusEffects;
-
-[RegisterComponent, NetworkedComponent]
-public sealed partial class CEPermanentStatusEffectsComponent : Component
-{
-    [DataField(required: true)]
-    public HashSet<EntProtoId> Effects = new();
-}
