+++ Content.Shared/_CE/StatusEffects/CERemoveStackOnHeal.cs
@@ -2,8 +2,25 @@ using Content.Shared._CE.Health;
 using Content.Shared._CE.StatusEffects.Core;
 using Content.Shared._CE.StatusEffects.Core.Components;
 using Content.Shared.StatusEffectNew;
+using Robust.Shared.GameStates;
 
-namespace Content.Shared._CE.StatusEffects.RemoveStackOnHeal;
+namespace Content.Shared._CE.StatusEffects;
+
+[RegisterComponent, NetworkedComponent]
+public sealed partial class CERemoveStackOnHealComponent : Component
+{
+    /// <summary>
+    /// How many heal units are required to remove 1 stack.
+    /// </summary>
+    [DataField]
+    public int HealPerStack = 1;
+
+    /// <summary>
+    /// If false, cant remove last status effect stack
+    /// </summary>
+    [DataField]
+    public bool CanRemoveStatusEffect = true;
+}
 
 public sealed class CERemoveStackOnHealSystem : EntitySystem
 {
