--- Content.Shared/_CE/StatusEffects/DamageVulnerability/CEDamageVulnerabilityStatusEffectComponent.cs
@@ -1,25 +1,0 @@
-using Content.Shared._CE.Health.Prototypes;
-using Robust.Shared.GameStates;
-using Robust.Shared.Prototypes;
-
-namespace Content.Shared._CE.DamageVulnerability;
-
-/// <summary>
-/// Status effect component that increases incoming damage of specific types.
-/// Each stack adds flat bonus damage of the configured type.
-/// </summary>
-[RegisterComponent, NetworkedComponent, AutoGenerateComponentState]
-public sealed partial class CEDamageVulnerabilityStatusEffectComponent : Component
-{
-    /// <summary>
-    /// How much extra damage each stack adds to incoming hits of the configured type.
-    /// </summary>
-    [DataField, AutoNetworkedField]
-    public int ExtraDamagePerStack = 1;
-
-    /// <summary>
-    /// Which damage types this vulnerability amplifies.
-    /// </summary>
-    [DataField]
-    public HashSet<ProtoId<CEDamageTypePrototype>> AffectedTypes = new() { "Physical" };
-}
