--- Content.Shared/_CE/StatusEffects/DivineShield/CEDivineShieldStatusEffectComponent.cs
@@ -1,8 +1,0 @@
-using Robust.Shared.GameStates;
-
-namespace Content.Shared._CE.DivineShield;
-
-[RegisterComponent, NetworkedComponent]
-public sealed partial class CEDivineShieldStatusEffectComponent : Component
-{
-}
