--- Content.Shared/_CE/StatusEffects/TempShield/CETempShieldStatusEffectComponent.cs
@@ -1,33 +1,0 @@
-using Content.Shared._CE.Health.Prototypes;
-using Robust.Shared.GameStates;
-using Robust.Shared.Prototypes;
-
-namespace Content.Shared._CE.TempShield;
-
-/// <summary>
-/// Status effect component for temporary shields.
-/// Each stack absorbs a configurable amount of damage from specific damage types.
-/// Stacks passively decay over time via the status effect stack system.
-/// </summary>
-[RegisterComponent, NetworkedComponent, AutoGenerateComponentState]
-public sealed partial class CETempShieldStatusEffectComponent : Component
-{
-    /// <summary>
-    /// How much damage each stack absorbs before being consumed.
-    /// </summary>
-    [DataField, AutoNetworkedField]
-    public int AbsorbPerStack = 1;
-
-    /// <summary>
-    /// Which damage types this shield absorbs.
-    /// Ignored when <see cref="All"/> is true.
-    /// </summary>
-    [DataField]
-    public HashSet<ProtoId<CEDamageTypePrototype>> AbsorbedTypes = new() { "Physical" };
-
-    /// <summary>
-    /// If true, absorbs a stack for any damage type, ignoring <see cref="AbsorbedTypes"/>.
-    /// </summary>
-    [DataField, AutoNetworkedField]
-    public bool All = false;
-}
