--- Content.Shared/_CE/StatusEffects/ToggleableStatusEffectAction/CETogglableActionComponent.cs
@@ -1,10 +1,0 @@
-using Robust.Shared.GameStates;
-
-namespace Content.Shared._CE.StatusEffects.ToggleableStatusEffectAction;
-
-[RegisterComponent, NetworkedComponent, AutoGenerateComponentState]
-public sealed partial class CEToggleableActionComponent : Component
-{
-    [DataField, AutoNetworkedField]
-    public bool Active;
-}
