+++ Resources/Prototypes/_CE/Skills/Knight/Passives/landing_technique.yml
@@ -2,7 +2,7 @@
   parent: BaseKnightSkill
   id: LandingTechnique
   icon:
-    sprite: _CE/Skills/todo.rsi
+    sprite: _CE/Skills/Knight/landing_technique.rsi
     state: icon
   effect: !type:AddStatusEffect
     effect: CEStatusEffectLandingTechnique
@@ -22,5 +22,5 @@
 - type: alert
   id: CELandingTechnique
   icons:
-  - sprite: _CE/Skills/todo.rsi
+  - sprite: _CE/Skills/Knight/landing_technique.rsi
     state: icon
