+++ Resources/Textures/_CE/Skills/Knight/landing_techique.rsi/meta.json
@@ -1,0 +1,14 @@
+{
+  "version": 1,
+  "size": {
+    "x": 32,
+    "y": 32
+  },
+  "license": "CC-BY-SA-4.0",
+  "copyright": "Created by TheShuEd",
+  "states": [
+    {
+      "name": "icon"
+    }
+  ]
+}
